Refactor the bal execution logic and add record support
