+++ gradle.properties
@@ -20,3 +20,4 @@ handlebarsVersion=4.5.0
 commonsCompressVersion=1.26.1
 asmVersion=9.2
 moduleWso2MiVersion=0.4.0-20251205-150600-4f43b8e
+miConnectorCoreVersion=0.9.3
