+++ native/src/main/java/io/ballerina/stdlib/mi/BalConnectorConfig.java
@@ -19,45 +19,34 @@
 package io.ballerina.stdlib.mi;
 import io.ballerina.runtime.api.Module;
 import io.ballerina.runtime.api.Runtime;
-import io.ballerina.runtime.api.creators.TypeCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
-import io.ballerina.runtime.api.types.Type;
-import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BError;
-import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
-import io.ballerina.runtime.api.values.BString;
-import org.apache.commons.collections.map.HashedMap;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.mediators.template.TemplateContext;
-import org.ballerinalang.langlib.value.FromJsonStringWithType;
 import org.wso2.integration.connector.core.AbstractConnector;
 import org.wso2.integration.connector.core.ConnectException;
 import org.wso2.integration.connector.core.connection.ConnectionHandler;
 
 import java.util.Stack;
 
-import static io.ballerina.stdlib.mi.Constants.*;
-
 public class BalConnectorConfig extends AbstractConnector {
     private static volatile Runtime rt = null;
     private static Module module = null;
+    private String orgName;
+    private String moduleName;
+    private String version;
+    private final BalExecutor balExecutor = new BalExecutor();
 
     public BalConnectorConfig() {
-        if (rt == null) {
-            synchronized (BalConnectorConfig.class) {
-                if (rt == null) {
-                    System.out.println("runtime created");
-                    ModuleInfo moduleInfo = new ModuleInfo();
-                    init(moduleInfo);
-                }
-            }
-        }
     }
 
     // This constructor is added to test the connector
     public BalConnectorConfig(ModuleInfo moduleInfo) {
-        init(moduleInfo);
+        this.orgName = moduleInfo.getOrgName();
+        this.moduleName = moduleInfo.getModuleName();
+        this.version = moduleInfo.getModuleVersion();
+        init();
     }
 
     @Override
@@ -71,8 +60,16 @@ public class BalConnectorConfig extends AbstractConnector {
          */
         // get name parameter value
         //TODO: Make any updated to support multiple clients
+        if (rt == null) {
+            synchronized (BalConnectorConfig.class) {
+                if (rt == null) {
+                    init();
+                }
+            }
+        }
         String connectorName = module.getName();
         String connectionName = lookupTemplateParamater(messageContext, "name");
+        String connectionType = lookupTemplateParamater(messageContext, "connectionType");
         //TODO: read parameter args and default values
         //TODO: Use the NodeParser API to generate the argument values
 
@@ -89,13 +86,13 @@ public class BalConnectorConfig extends AbstractConnector {
 //                Type type = recValue.getType();
 //                Object o = FromJsonStringWithType.fromJsonStringWithType(jsonString, ValueCreator.createTypedescValue(type));
 
-                Object[] args = new Object[Integer.parseInt(messageContext.getProperty(Constants.SIZE).toString())];
-                setParameters(args, messageContext);
-                clientObject = ValueCreator.createObjectValue(module, messageContext.getProperty("objectTypeName").toString(), args);
+                Object[] args = new Object[Integer.parseInt(messageContext.getProperty(connectionType + "_" + Constants.SIZE).toString())];
+                setParameters(args, messageContext, connectionType);
+                clientObject = ValueCreator.createObjectValue(module, messageContext.getProperty(connectionType + "_objectTypeName").toString(), args);
             } catch (BError clientError) {
                 handleException(clientError.getMessage(), messageContext);
             }
-            BalConnectorConnection balConnection = new BalConnectorConnection(module, messageContext.getProperty("objectTypeName").toString(), clientObject);
+            BalConnectorConnection balConnection = new BalConnectorConnection(module, messageContext.getProperty(connectionType + "_objectTypeName").toString(), clientObject);
             try {
                 handler.createConnection(connectorName, connectionName, balConnection, messageContext);
             } catch (NoSuchMethodError e) {
@@ -105,8 +102,8 @@ public class BalConnectorConfig extends AbstractConnector {
         messageContext.setProperty("connectionName", connectionName);
         }
 
-    private void init (ModuleInfo moduleInfo){
-        module = new Module(moduleInfo.getOrgName(), moduleInfo.getModuleName(), moduleInfo.getModuleVersion());
+    private void init() {
+        module = new Module(orgName, moduleName, version);
         rt = Runtime.from(module);
         rt.init();
         rt.start();
@@ -126,9 +123,9 @@ public class BalConnectorConfig extends AbstractConnector {
         return currentFuncHolder.getParameterValue(paramName).toString();
     }
 
-    private void setParameters(Object[] args, MessageContext context) {
+    private void setParameters(Object[] args, MessageContext context, String connectionType) {
         for (int i = 0; i < args.length; i++) {
-            Object param = getParameter(context, "param" + i, "paramType" + i, i);
+            Object param = balExecutor.getParameter(context, connectionType + "_param" + i, connectionType + "_paramType" + i, i);
             //TODO: check handling null parameters
 //            if (param == null) {
 //                return false;
@@ -137,32 +134,27 @@ public class BalConnectorConfig extends AbstractConnector {
         }
     }
 
-    private Object getParameter(MessageContext context, String value, String type, int index) {
-        String paramName = context.getProperty(value).toString();
-        Object param = lookupTemplateParamater(context, paramName);
-        String stringParam = ((String) param).replaceAll("^\"(.*)\"$", "$1");
-        //TODO: check handling null parameters
-//        if (param == null) {
-//            log.error("Error in getting the ballerina function parameter: " + paramName);
-//            return null;
-//        }
-        String paramType = context.getProperty(type).toString();
-        return switch (paramType) {
-            //TODO: Revisit handling union and record types
-            case BOOLEAN -> Boolean.parseBoolean(stringParam);
-            case INT -> Long.parseLong(stringParam);
-            case STRING, UNION -> StringUtils.fromString(stringParam);
-            case FLOAT -> Double.parseDouble(stringParam);
-            case DECIMAL -> ValueCreator.createDecimalValue(stringParam);
-            case RECORD -> createRecordValue(stringParam, context, index);
-            default -> null;
-        };
+    public String getOrgName() {
+        return orgName;
     }
 
-    private Object createRecordValue(String jsonString, MessageContext context, int paramIndex) {
-        String recordName = context.getProperty("recordName" + paramIndex).toString();
-        BMap<BString, Object> recValue = ValueCreator.createRecordValue(BalConnectorConfig.getModule(), recordName);
-        Type recType = recValue.getType();
-        return FromJsonStringWithType.fromJsonStringWithType(StringUtils.fromString(jsonString), ValueCreator.createTypedescValue(recType));
+    public void setOrgName(String orgName) {
+        this.orgName = orgName;
+    }
+
+    public String getModuleName() {
+        return moduleName;
+    }
+
+    public void setModuleName(String moduleName) {
+        this.moduleName = moduleName;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
     }
 }
