+++ native/src/main/java/io/ballerina/stdlib/mi/Mediator.java
@@ -20,48 +20,18 @@ package io.ballerina.stdlib.mi;
 
 import io.ballerina.runtime.api.Module;
 import io.ballerina.runtime.api.Runtime;
-import io.ballerina.runtime.api.creators.ValueCreator;
-import io.ballerina.runtime.api.utils.JsonUtils;
-import io.ballerina.runtime.api.utils.StringUtils;
-import io.ballerina.runtime.api.values.BArray;
-import io.ballerina.runtime.api.values.BDecimal;
-import io.ballerina.runtime.api.values.BError;
-import io.ballerina.runtime.api.values.BMap;
-import io.ballerina.runtime.api.values.BString;
-import io.ballerina.runtime.api.values.BXml;
-import org.apache.axiom.om.OMElement;
-import org.apache.axiom.om.util.AXIOMUtil;
 import org.apache.axis2.AxisFault;
-import org.apache.synapse.data.connector.ConnectorResponse;
-import org.apache.synapse.data.connector.DefaultConnectorResponse;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.mediators.AbstractMediator;
-import org.apache.synapse.mediators.elementary.EnrichMediator;
-import org.apache.synapse.mediators.elementary.Source;
-import org.apache.synapse.mediators.elementary.Target;
-import org.apache.synapse.mediators.template.TemplateContext;
-import org.apache.synapse.util.MediatorEnrichUtil;
-
-import java.util.Objects;
-import java.util.Stack;
-
-import static io.ballerina.stdlib.mi.Constants.ARRAY;
-import static io.ballerina.stdlib.mi.Constants.BOOLEAN;
-import static io.ballerina.stdlib.mi.Constants.DECIMAL;
-import static io.ballerina.stdlib.mi.Constants.FLOAT;
-import static io.ballerina.stdlib.mi.Constants.INT;
-import static io.ballerina.stdlib.mi.Constants.JSON;
-import static io.ballerina.stdlib.mi.Constants.STRING;
-import static io.ballerina.stdlib.mi.Constants.XML;
 
 public class Mediator extends AbstractMediator {
 
-    private static final String TEMP_RESPONSE_PROPERTY_NAME = "TEMP_BAL_RESPONSE_PROPERTY_";
     private static volatile Runtime rt = null;
     private static Module module = null;
     private String orgName;
     private String moduleName;
     private String version;
+    private final BalExecutor balExecutor = new BalExecutor();
 
     public Mediator() {
     }
@@ -74,14 +44,6 @@ public class Mediator extends AbstractMediator {
         init();
     }
 
-    private static String getResultProperty(MessageContext context) {
-        return lookupTemplateParameter(context, Constants.RESPONSE_VARIABLE).toString();
-    }
-
-    private static boolean isOverwriteBody(MessageContext context) {
-        return Boolean.parseBoolean((String) lookupTemplateParameter(context, Constants.OVERWRITE_BODY));
-    }
-
     public boolean mediate(MessageContext context) {
 
         if (rt == null) {
@@ -91,160 +53,12 @@ public class Mediator extends AbstractMediator {
                 }
             }
         }
-        String balFunctionReturnType = context.getProperty(Constants.RETURN_TYPE).toString();
-        Object[] args = new Object[Integer.parseInt(context.getProperty(Constants.SIZE).toString())];
-        if (!setParameters(args, context)) {
-            return false;
-        }
         try {
-            Object result = rt.callFunction(module, context.getProperty(Constants.FUNCTION_NAME).toString(), null, args);
-            if (Objects.equals(balFunctionReturnType, XML)) {
-                result = BXmlConverter.toOMElement((BXml) result);
-            } else if (Objects.equals(balFunctionReturnType, DECIMAL)) {
-                result = ((BDecimal) result).value().toString();
-            } else if (Objects.equals(balFunctionReturnType, STRING)) {
-                result = ((BString) result).getValue();
-            } else if (Objects.equals(balFunctionReturnType, ARRAY) || result instanceof BArray) {
-                // Convert BArray to JSON string format for MI consumption
-                result = TypeConverter.arrayToJsonString((BArray) result);
-            } else if (result instanceof BMap) {
-                result = result.toString();
-            }
-            ConnectorResponse connectorResponse = new DefaultConnectorResponse();
-            if (isOverwriteBody(context)) {
-                context.setProperty(TEMP_RESPONSE_PROPERTY_NAME + getResultProperty(context), result);
-                overwriteBody(context, result);
-            } else {
-                connectorResponse.setPayload(result);
-            }
-            context.setVariable(getResultProperty(context), connectorResponse);
-        } catch (BError bError) {
-            handleException(bError.getMessage(), context);
+            return balExecutor.execute(rt, module, context);
         } catch (AxisFault e) {
-            handleException("Error while overwriting message body", e, context);
+            handleException("Error while executing ballerina", e, context);
         }
-        return true;
-    }
-
-    protected void overwriteBody(MessageContext messageContext, Object payload) throws AxisFault {
-        Source source = MediatorEnrichUtil.createSourceWithProperty(TEMP_RESPONSE_PROPERTY_NAME + getResultProperty(messageContext));
-        Target target = MediatorEnrichUtil.createTargetWithBody();
-        doEnrich(messageContext, source, target);
-    }
-
-    private void doEnrich(MessageContext synCtx, Source source, Target target) {
-        if (!Boolean.TRUE.equals(synCtx.getProperty("message.builder.invoked"))) {
-            MediatorEnrichUtil.buildMessage(synCtx);
-        }
-
-        EnrichMediator enrichMediator = new EnrichMediator();
-        enrichMediator.setSource(source);
-        enrichMediator.setTarget(target);
-        enrichMediator.setNativeJsonSupportEnabled(true);
-        enrichMediator.mediate(synCtx);
-    }
-
-    private boolean setParameters(Object[] args, MessageContext context) {
-        for (int i = 0; i < args.length; i++) {
-            Object param = getParameter(context, "param" + i, "paramType" + i);
-            if (param == null) {
-                return false;
-            }
-            args[i] = param;
-        }
-        return true;
-    }
-
-    private Object getParameter(MessageContext context, String value, String type) {
-        String paramName = context.getProperty(value).toString();
-        Object param = lookupTemplateParameter(context, paramName);
-        if (param == null) {
-            log.error("Error in getting the ballerina function parameter: " + paramName);
-            return null;
-        }
-        String paramType = context.getProperty(type).toString();
-        return switch (paramType) {
-            case BOOLEAN -> Boolean.parseBoolean((String) param);
-            case INT -> Long.parseLong((String) param);
-            case STRING -> StringUtils.fromString((String) param);
-            case FLOAT -> Double.parseDouble((String) param);
-            case DECIMAL -> ValueCreator.createDecimalValue((String) param);
-            case JSON -> getBMapParameter(param);
-            case XML -> getBXmlParameter(context, value);
-            case ARRAY -> getArrayParameter((String) param, context, value);
-            default -> null;
-        };
-    }
-
-    private BXml getBXmlParameter(MessageContext context, String parameterName) {
-        OMElement omElement = getOMElement(context, parameterName);
-        if (omElement == null) {
-            return null;
-        }
-        return OMElementConverter.toBXml(omElement);
-    }
-
-    private OMElement getOMElement(MessageContext ctx, String value) {
-        String param = ctx.getProperty(value).toString();
-        Object paramValue = lookupTemplateParameter(ctx, param);
-        if (paramValue != null) {
-            if (paramValue instanceof OMElement) {
-                return (OMElement) paramValue;
-            } else {
-                try {
-                    return AXIOMUtil.stringToOM((String) lookupTemplateParameter(ctx, param));
-                } catch (Exception ignored) {
-                }
-            }
-        }
-        log.error("Error in getting the OMElement");
-        return null;
-    }
-
-    public static Object lookupTemplateParameter(MessageContext ctx, String paramName) {
-        Stack funcStack = (Stack) ctx.getProperty(Constants.SYNAPSE_FUNCTION_STACK);
-        TemplateContext currentFuncHolder = (TemplateContext) funcStack.peek();
-        return currentFuncHolder.getParameterValue(paramName);
-    }
-
-    private BMap getBMapParameter(Object param) {
-        if (param instanceof String) {
-            return (BMap) JsonUtils.parse((String) param);
-        } else {
-            return (BMap) JsonUtils.parse(param.toString());
-        }
-    }
-
-    /**
-     * Get array parameter from context and convert to Ballerina array.
-     * Delegates to TypeConverter for the actual conversion logic.
-     *
-     * @param jsonArrayString JSON array string
-     * @param context         Message context
-     * @param valueKey        The property key like "param0" used to extract the parameter
-     */
-    private Object getArrayParameter(String jsonArrayString, MessageContext context, String valueKey) {
-        // Extract parameter index from valueKey (e.g., "param0" -> 0)
-        int paramIndex = -1;
-        String indexStr = valueKey.replaceAll("\\D+", "");
-        if (!indexStr.isEmpty()) {
-            try {
-                paramIndex = Integer.parseInt(indexStr);
-            } catch (NumberFormatException e) {
-                log.error("Invalid parameter index in valueKey: " + valueKey, e);
-                // Optionally, handle error (e.g., throw, return null, etc.)
-                return null;
-            }
-        } else {
-            log.error("No digits found in valueKey: " + valueKey);
-            return null;
-        }
-
-        // Get the array element type from context
-        String elementType = context.getProperty("arrayElementType" + paramIndex).toString();
-
-        // Use shared TypeConverter for conversion
-        return TypeConverter.convertToArray(jsonArrayString, elementType);
+        return false;
     }
 
     private void init() {
