+++ tests/build.gradle
@@ -36,7 +36,7 @@ dependencies {
     testImplementation project(path: ':tool-mi-module-gen-cli')
 }
 
-def projects = ['project2', 'project3', 'arrayProject']
+def projects = ['project2', 'project3', 'arrayProject', 'recordProject']
 def toolVersion = stripBallerinaExtensionVersion("${project.version}")
 // Define balCommand to use the test-specific distribution
 ext.balCommand = Os.isFamily(Os.FAMILY_WINDOWS)
@@ -211,6 +211,9 @@ test {
 
     systemProperty "bal.command", balCommand.toString()
 
+    if (System.getProperty('debug') == 'true') {
+        jvmArgs '-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=*:5005'
+    }
     useTestNG() {
         suites 'src/test/resources/testng.xml'
     }
