+++ tool-mi-module-gen-cli/src/main/resources/balConnector/functions/functions_template.xml
@@ -33,6 +33,9 @@
         {{#each functionParams ~}}
         <property name="param{{@index}}" value="{{value}}"/>
         <property name="paramType{{@index}}" value="{{paramType}}"/>
+        {{#eq paramType "record" ~}}
+        <property name="param{{@index}}_recordName" value="{{unwrapOptional typeSymbol.name}}" />
+        {{/eq~}}
         {{/each ~}}
         {{#each params ~}}
         <property name="{{key}}" value="{{value}}"/>
